Firebase generates all cities at lobby

+++ Stratagem/Game Scenes/Helpers/Updater.swift
@@ -6,7 +6,7 @@ class Updater: SKScene {
     var lastTime: Float?
     override func update(_ currentTime: TimeInterval) {
         if lastTime != nil {
-            Global.gameVars?.update(deltaTime: Float(currentTime) - lastTime!)
+            Global.gameVars?.update(deltaTime: CGFloat(currentTime) - CGFloat(lastTime!))
             lastTime = Float(currentTime)
         }
         else {lastTime = Float(currentTime)}
